+++ src/a2x_pack_state.c
@@ -236,7 +236,7 @@ void a_state_exit(void)
 {
     a_out__state("Telling all states to exit");
 
-    for(size_t i = a_list_size(g_stack); i--; ) {
+    for(unsigned i = a_list_size(g_stack); i--; ) {
         pending_new(A_STATE_ACTION_POP, NULL);
     }
 }
